new file checkIfEqual
