+++ main/src/js/preparePage.js
@@ -136,63 +136,6 @@ function getFAppFromId(id) {
   }
 }
 
-function checkIfEqual(id, a, b, dsList) {
-  var equ = a + '=' + b;
-  checkIfEquality(id, equ, dsList);
-}
-
-function checkIfEquality(id, equ, dsList) {
-  console.log(equ);
-  var myTree = parse(equ);
-  myTree = fillWithRandomValAndCheckDefSets(myTree, dsList);
-  var almostOne = evaluateTree(myTree);
-  var dif = Math.abs(almostOne - 1);
-  var fApp = getFAppFromId(id);
-  var precision = fApp.precision;
-  if (dif < precision) {
-    $('#' + id).removeClass('mod_wrong').addClass('mod_ok');
-  } else {
-    $('#' + id).removeClass('mod_ok').addClass('mod_wrong');
-  }
-}
-
-function fillWithRandomValAndCheckDefSets(treeVar, dsList) {
-  var rememberTree = JSON.stringify(treeVar);
-  if (dsList.length == 0) {
-    fillWithValues(treeVar);
-    return treeVar;
-  } else {
-    // start watchdog
-    var success = false;
-    var start = new Date();
-    var timePassedMilliseconds = 0;
-    while (!success && timePassedMilliseconds < 2000) {
-      var tree2 = new FaTree();
-      tree2 = JSON.parse(rememberTree);
-      fillWithValues(tree2);
-      var variableValueList = tree2.variableValueList;
-      // CheckDefinitionSets
-      for (var i = 0; i < dsList.length; i++) {
-        var definitionset = parse(dsList[i]);
-        fillWithValues(definitionset, variableValueList);
-        var value = evaluateTree(definitionset);
-        success = ((value > 0) || typeof value == 'undefined');
-        if (!success) {
-          // short circuit
-          i = dsList.length;
-          // restore leafs with value = undefined
-        }
-      }
-      var now = new Date();
-      timePassedMilliseconds = now.getTime() - start.getTime();
-    }
-    if (!success) {
-      tree2.hasValue = false;
-      tree2.variableValueList = [];
-    }
-    return tree2;
-  }
-}
 
 function makeAutoUnitstring(mf) {
   // mf = MathField
